asim: move pacer into queue pkg
This commit moves the replica pacer to the `asim/queue` pkg. The pacer
is used to add replicas into the queues at some rate. Moving it out of
the base `asim` pkg clears up the pacer relation to the rest of the
simulation code.

Release note: None
