+++ pkg/kv/kvserver/asim/BUILD.bazel
@@ -3,10 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
 
 go_library(
     name = "asim",
-    srcs = [
-        "asim.go",
-        "pacer.go",
-    ],
+    srcs = ["asim.go"],
     importpath = "github.com/cockroachdb/cockroach/pkg/kv/kvserver/asim",
     visibility = ["//visibility:public"],
     deps = [
@@ -24,13 +21,10 @@ go_library(
 
 go_test(
     name = "asim_test",
-    srcs = [
-        "asim_test.go",
-        "pacer_test.go",
-    ],
+    srcs = ["asim_test.go"],
     args = ["-test.timeout=295s"],
-    embed = [":asim"],
     deps = [
+        ":asim",
         "//pkg/kv/kvserver/asim/config",
         "//pkg/kv/kvserver/asim/metrics",
         "//pkg/kv/kvserver/asim/state",
